requirements.txt
python-telegram-bot 
yfinance 
schedule